externalisation des fonctions
